feat(x): allow timeline embedding via url
